change long to int

+++ src/main/java/com/pbl/pbl_be/exception/ResourceNotFoundException.java
@@ -4,9 +4,9 @@ package com.pbl.pbl_be.exception;
 public class ResourceNotFoundException extends RuntimeException{
     String resourceName;
     String fieldName;
-    long fieldValue;
+    int fieldValue;
 
-    public ResourceNotFoundException(String resourceName, String fieldName, long fieldValue) {
+    public ResourceNotFoundException(String resourceName, String fieldName, int fieldValue) {
         super(String.format("%s not found with %s : %s",resourceName,fieldName,fieldValue));
         this.resourceName = resourceName;
         this.fieldName = fieldName;
